api/requirements.txt
autopep8==1.5.4
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
cassandra-driver==3.24.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
geomet==0.2.1.post1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
oauthlib==3.1.0
Pillow==8.0.1
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
